Update SenderX.cpp
Edited to implement crc if crcflg is true, else do checksum.
Because crc is implemented, size of blockbuffer is 133 bytes, not 132 (due to crc being 2 bytes), so edited indices of buffer in response to size change. Also padded ctrl-z onto blkbuf[1] in case of checksum, because checksum is only 1 byte.